# src/main/resources/openapi/aequitas-api.yaml
openapi: 3.0.3
info:
  title: Aequitas Central Service API
  description: |
    REST API for managing time entries and user profiles within a multi-tenant legal timekeeping SaaS platform.
    
    This API follows hexagonal architecture principles and enforces tenant-level data isolation.
    All operations are scoped to the authenticated tenant context.
    
    ## Authentication
    All endpoints require authentication via OAuth2/JWT tokens. The security context provides
    tenant identification and role-based access control.
    
    ## Idempotency
    Mutation endpoints support optional `Idempotency-Key` headers to prevent duplicate operations
    caused by client retries. Keys should be UUIDs and are scoped to tenant + operation type.
    
    ## Pagination
    List endpoints use cursor-based pagination for stable iteration over large result sets.
  version: 0.1.0
  contact:
    name: Aequitas Engineering
    url: https://example.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://api.aequitas.example.com
    description: Production server
...
paths:
  /api/v1/auth/signup:
    post:
      tags:
        - Authentication
      summary: Register a new user
      description: |
        Creates a new user account in Supabase with firm association and role assignment.
        
        This endpoint creates a new user account with the provided email and password, then links the user
        to the specified firm identifier and assigns the requested role. Upon successful registration,
        Supabase generates session tokens (access and refresh) which are returned in the response.
      operationId: signUp
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignUpRequest'
      responses:
        '201':
          description: User successfully registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Invalid request payload (validation failure)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Firm not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: User with this email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Supabase service error or internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/auth/signin:
    post:
      tags:
        - Authentication
      summary: Authenticate an existing user
      description: |
        Validates email and password credentials and returns a new session with access tokens.
        
        This endpoint validates the provided email and password against Supabase and, if successful,
        returns a new session containing fresh access and refresh tokens. The access token can be used
        for subsequent API calls requiring authentication.
      operationId: signIn
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignInRequest'
      responses:
        '200':
          description: User successfully authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Invalid request payload (validation failure)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Invalid credentials (wrong email or password)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Supabase service error or internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    SignUpRequest:
      type: object
      required:
        - firmId
        - email
        - password
        - role
      properties:
        firmId:
          type: string
          format: uuid
          description: Firm/tenant identifier to associate the user with
        email:
          type: string
          format: email
          description: User's email address (RFC 5322 compliant)
        password:
          type: string
          minLength: 8
          description: User's password (minimum 8 characters)
        role:
          $ref: '#/components/schemas/Role'
      example:
        firmId: "f47ac10b-58cc-4372-a567-0e02b2c3d479"
        email: "user@example.com"
        password: "SecurePass123!"
        role: "EMPLOYEE"
    SignInRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          description: User's email address
        password:
          type: string
          description: User's password
      example:
        email: "user@example.com"
        password: "SecurePass123!"
    AuthResponse:
      type: object
      required:
        - user
        - tokens
      properties:
        user:
          $ref: '#/components/schemas/UserProfileResponse'
        tokens:
          $ref: '#/components/schemas/TokenResponse'
      description: Authentication response combining user profile and bearer tokens
      example:
        user:
          id: "550e8400-e29b-41d4-a716-446655440000"
          firmId: "660e8400-e29b-41d4-a716-446655440001"
          email: "user@example.com"
          role: "EMPLOYEE"
        tokens:
          accessToken: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          refreshToken: "v1.MRjJJojzaRO3oUK..."
          expiresInSeconds: 3600
          tokenType: "bearer"
    TokenResponse:
      type: object
      required:
        - accessToken
        - refreshToken
        - expiresInSeconds
        - tokenType
      properties:
        accessToken:
          type: string
          description: Short-lived JWT access token
        refreshToken:
          type: string
          description: Token used to refresh the access token
        expiresInSeconds:
          type: integer
          format: int64
          description: Seconds until the access token expires
        tokenType:
          type: string
          description: Token type prefix (e.g., 'bearer')
          example: "bearer"
      example:
        accessToken: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        refreshToken: "v1.MRjJJojzaRO3oUK..."
        expiresInSeconds: 3600
        tokenType: "bearer"
...
    UserProfileResponse:
      type: object
      required:
        - id
        - firmId
        - email
        - role
      properties:
        id:
          type: string
          format: uuid
          description: User identifier
        firmId:
          type: string
          format: uuid
          description: Tenant identifier
        email:
          type: string
          format: email
          description: Email address
        role:
          $ref: '#/components/schemas/Role'
      example:
        id: "550e8400-e29b-41d4-a716-446655440000"
        firmId: "660e8400-e29b-41d4-a716-446655440001"
        email: "john.doe@example.com"
        role: "EMPLOYEE"
...
    Role:
      type: string
      enum:
        - EMPLOYEE
        - MANAGER
        - ADMIN
      description: |
        RBAC role within a firm:
        * `EMPLOYEE` - Individual contributor with access to own entries only
        * `MANAGER` - People lead who may manage subordinate time entries
        * `ADMIN` - Firm administrator with full access inside the tenant
    ErrorResponse:
      type: object
      required:
        - timestamp
        - status
        - error
        - message
        - path
      properties:
        timestamp:
          type: string
          format: date-time
          description: Error occurrence timestamp
        status:
          type: integer
          description: HTTP status code
        error:
          type: string
          description: HTTP status message
        message:
          type: string
          description: Detailed error message
        path:
          type: string
          description: Request path that caused the error
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
          description: Field-level validation errors (for 400 responses)
      example:
        timestamp: "2025-11-15T10:30:00Z"
        status: 400
        error: "Bad Request"
        message: "Validation failed"
        path: "/api/v1/entries"
        errors:
          - field: "narrative"
            message: "must not be blank"
  securitySchemes:
    oauth2:
      type: oauth2
      description: OAuth2 authentication with JWT tokens
      flows:
        authorizationCode:
          authorizationUrl: https://auth.aequitas.example.com/oauth/authorize
          tokenUrl: https://auth.aequitas.example.com/oauth/token
          scopes:
            read: Read access to resources
            write: Write access to resources
security:
  - oauth2: []